search to minimize expected travel time

+++ CBS-corridor/CBSH-rect-cmake/LLNode.cpp
@@ -2,7 +2,7 @@
 
 LLNode::LLNode() : loc(0), g_val(0), h_val(0), parent(nullptr), timestep(0) {}
 
-LLNode::LLNode(int loc, int g_val, int h_val, LLNode* parent, int timestep) :
+LLNode::LLNode(int loc, float g_val, int h_val, LLNode* parent, int timestep) :
         loc(loc), g_val(g_val), h_val(h_val), parent(parent), timestep(timestep) {}
 
 //LLNode::LLNode(const LLNode& other)
